Finalized ADC, DAC functions
ADC to percent still needs to be scaled to percent. Equation is currently missing. Drive_by_Wire.c will need to be updated to reflect DAC instead of digital potentiometers.

+++ Throttle/dac.h
@@ -12,4 +12,4 @@
 #include "driverlib/i2c.h"
 #include "driverlib/pin_map.h"
 
-uint16_t SetDACVoltage(uint16_t setPoint);
+uint32_t SetDACVoltage(uint32_t setPoint);
